posgresTest/pyproject.toml
[project]
name = "posgrestest"
version = "0.1.0"
description = ""
authors = [
    {name = "Yami",email = "renard.ryan74@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
